Hack together an ESM friendly-ish build
The esbuild.js fixes the runtime dependence of using require()
throughout, so that we can use this in a strictly ESM environment. It is
a complete hack, but quickly solved my problem.

I did NOT get the index.d.ts file tuned correctly so that `parse` and
`render` could be imported properly. Instead, in the typescript code
that consumes these, `as` casts need to be used. Again, just a
workaraound, but got it working for now.